% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.R
\name{adaptive_perm_test}
\alias{adaptive_perm_test}
\title{Adaptive permutation test}
\usage{
adaptive_perm_test(x1, x2, g1, g2, stat, B, alpha = 0.025, x3 = NULL,
  g3 = NULL)
}
\arguments{
\item{x1}{First stage observations}

\item{x2}{Second stage observations}

\item{g1}{First stage treatment assignments}

\item{g2}{Second stage treatment assignments (see Details)}

\item{stat}{Function that computes the test statistic}

\item{B}{Number of draws from the permutation distribution to use}

\item{alpha}{Significance level of the test}

\item{x3}{Third stage observations (e.g. sample size increase)}

\item{g3}{Third stage treatment allocations (e.g. sample size increase)}
}
\value{
logical indicating wether the test rejects
}
\description{
Perform an adaptive permutation test
}
\details{
Note that we currently assume that the group sizes of the second
stage in the adapted trial are equal to those of the preplanned
test.
}
\author{
Florian Klinglmueller
}

