% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{adaptive_permtest_2s}
\alias{adaptive_permtest_2s}
\title{Compare two-sample adaptive tests}
\usage{
adaptive_permtest_2s(x, y, n1, n, ne, m1, m, me, test_statistic,
  perms = 50000, alpha = 0.025)
}
\arguments{
\item{n1}{first stage sample size (control group)}

\item{n}{preplanned total sample size (control group)}
}
\description{
Compare adaptive test procedures for general two-sample cases
}
\details{
\code{rdist} needs to take the number of samples to return as its first argument.
}

