% Generated by roxygen2 (4.0.0): do not edit by hand
\name{adaptive_ttest_os}
\alias{adaptive_ttest_os}
\title{Adaptive t-test}
\usage{
adaptive_ttest_os(x, n1, n, ne, alpha = 0.025, mpfr = FALSE)
}
\arguments{
\item{mpfr}{Whether to use high precision numbers from \code{Rmpfr}}

\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{alpha}{}
}
\value{
Logical whether the test rejects
}
\description{
Adaptive t-test as described in Timmesfeld et al. (2007)
}
\details{
Warning the current implementation seems to be numerically instable
}
\author{
Florian Klinglmueller
}

