% Generated by roxygen2 (4.0.0): do not edit by hand
\name{all_assignments}
\alias{all_assignments}
\title{All assignments}
\usage{
all_assignments(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
logical matrix
}
\description{
Returns a logical matrix with all possible assignments (e.g. of patients) of \code{n} objects to two groups (e.g. treatment and control)
}
\author{
Florian Klinglmueller
}

