% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{all_reassignments}
\alias{all_reassignments}
\title{All combinations of two group assignments}
\usage{
all_reassignments(n, k)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of observations in (e.g. treatment) group}
}
\value{
integer matrix of size \code{n} x \code{choose(n,k)}
}
\description{
Returns a logical matrix with all possible assignments (e.g. of patients) of \code{n} objects to two groups (e.g. treatment and control) given that \code{k} objects are in one group
}
\author{
Florian Klinglmueller
}

