% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timmesfeld.R
\name{clev}
\alias{clev}
\title{Condition level function adaptive t-test}
\usage{
clev(tu2, u1, v1, tn2, n, n1, alpha = 0.025, mpfr = F)
}
\arguments{
\item{tu2}{Sum of squares of extended second stage}

\item{u1}{First stage sum of squares}

\item{v1}{First stage sum of observations}

\item{tn2}{Extended second stage sample size}

\item{n}{Preplanned total sample size}

\item{n1}{First stage sample size}

\item{alpha}{Significance level}
}
\value{
Conditional level for the adapted test
}
\description{
Computes the conditional level of the adaptive one-sided one-sample t-test for the null hypothesis mu=0 against the alternative mu > 0, given first stage sum of observations, sum of squares, and adapted secon stage sum of squares
}
\examples{
## From Timmesfeld et al. (2007) - should be 0.042
tu2 = 49968.5
u1 = 7989
v1 = 82.5
tn2 = 104
n1 = 15
n = 65
clev(tu2,u1,v1,tn2,n,n1)
tccv(tu2,u1,v1,tn2,n,n1)
}

