% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timmesfeld.R
\name{clev_twosample}
\alias{clev_twosample}
\title{Conditional level of extended two sample t-test}
\usage{
clev_twosample(tuu2, uu1, v1, tm2, tn2, m, n, m1, n1, alpha = 0.025)
}
\arguments{
\item{tuu2}{Two dimensional vector containing the sum and the sum of squares of the extended second stage observations}

\item{uu1}{Two dimensional vector containing the  sum and the sum of squares of the first stage observations}

\item{v1}{First stage sum of observations}

\item{tm2}{Extended second stage sample size treatment group}

\item{tn2}{Extended second stage sample size}

\item{m}{Preplanned total sample size treatment group}

\item{n}{Preplanned total sample size}

\item{m1}{First stage sample size treatment group}

\item{n1}{First stage sample size}

\item{alpha}{Significance level}

\item{tu2}{Sum of squares of extended second stage}

\item{u1}{First stage sum of squares}

\item{v1}{Sum of first stage treatment group observations}

\item{tn2}{Extended second stage sample size control group}

\item{n}{Preplanned total sample size control group}

\item{n1}{First stage sample size control group}

\item{alpha}{Significance level}
}
\value{
Conditional level of an extended two sample t-test
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\examples{
## From Timmesfeld et al. (2007) - should be 0.042
tu2 = 49968.5
u1 = 7989
v1 = 82.5
tn2 = 104
n1 = 15
n = 65
clev(tu2,u1,v1,tn2,n,n1)
tccv(tu2,u1,v1,tn2,n,n1)
}
\author{
Florian Klinglmueller
}

