% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_test.R
\name{closed_pval}
\alias{closed_pval}
\title{adjusted closed test pvalue}
\usage{
closed_pval(i, p, m = log2(length(p) + 1))
}
\arguments{
\item{i}{elementary hypothesis to test}

\item{p}{list of local p-values (see details)}

\item{m}{overall number of elementary hypotheses}
}
\value{
adjusted p-value
}
\description{
compute ajusted p-value for an elementary hypothesis based on a closed test procedure
}
\author{
Florian Klinglmueller
}

