% Generated by roxygen2 (4.0.0): do not edit by hand
\name{closed_test}
\alias{closed_test}
\title{closed test decision}
\usage{
closed_test(i, p, m = log2(length(p) + 1), alpha = 0.025)
}
\arguments{
\item{i}{elementary hypothesis to test}

\item{p}{list of local p-values (see details)}

\item{alpha}{significance level}
}
\value{
logical whether hypothesis can be rejected at FWER level \code{alpha}
}
\description{
compute test decision for an elementary hypothesis based on a closed test procedure
}
\author{
Florian Klinglmueller
}

