% Generated by roxygen2 (4.0.0): do not edit by hand
\name{closure}
\alias{closure}
\title{Perform a closed test procedure}
\usage{
closure(p, A)
}
\arguments{
\item{p}{vector of unadjusted p-values}

\item{A}{matrix of local critical levels sorted by binary representation of intersection hypotheses (see details)}
}
\value{
logical vector wether elementary hypotheses can be rejected
}
\description{
Performs a closed test procedure using a binary representation of intersection hypotheses
}
\details{
Intersection hypothesis - of a subset of m overall elementary hypotheses - can be represented by integer numbers of length m. \code{closure} takes a matrix of local significance levels ordered by the value of the binary representation of the corresponding intersection hypothesis.
}
\author{
Florian Klinglmueller
}

