% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_power}
\alias{compare_power}
\title{Type 1 error}
\usage{
compare_power(X = rnorm, X2 = NULL, ncp, n, n1, stat = meandiff, B, MCMC,
  one_sample = FALSE, ...)
}
\arguments{
\item{X}{Random number generater for first stage observations}

\item{X2}{Random number generator for second stage observations (if NULL first stage generator will be used)}

\item{ncp}{Non-centrality (shift) paramater}

\item{n}{pre-planned total sample size}

\item{n1}{first stagen sample size}

\item{stat}{}

\item{B}{number of permutations to be used to compute the permutation null distribution}

\item{MCMC}{number of simulation runs for Type I error estimation}

\item{one_sample}{should a one-sample test be simulated}

\item{...}{additional functions to conditional error rate function}
}
\value{
matrix of p-values from different test procedures
}
\description{
Template function to compute the Type 1 error of a trial design
}
\author{
Florian Klinglmueller
}

