% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{cond_power_rule_norm}
\alias{cond_power_rule_norm}
\title{Conditional power sample size reassessment rule (z-test)}
\usage{
cond_power_rule_norm(x1, m = 1, target = 0.9, alpha = 0.025, maxN = Inf)
}
\arguments{
\item{x1}{First stage observations}

\item{m}{Target mean value}

\item{target}{Target power}

\item{alpha}{Significance level}

\item{maxN}{Maximum sample size}
}
\value{
Total sample size required to achieve the target power
}
\description{
Conditional power rule for the one-sample (or paired) z-test using the normal distribution sample size formula. Reestimates the standard deviation from the first stage and recomputes the sample size such that the power to reject the null meets the target power assuming that the mean (paired treatment difference) is equal to a prespecified value.
}
\author{
Florian Klinglmueller
}

