% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closed_test.R
\name{contains}
\alias{contains}
\title{Contains i}
\usage{
contains(i, m)
}
\arguments{
\item{i}{position}

\item{m}{number of digits}
}
\value{
integer vector
}
\description{
Computes and returns all numbers that as binary numbers of length \code{m} that have a 1 at the \code{i}-th position. Can be used to represent intersection hypotheses for efficient implementation of closed testing procedures.
}
\examples{
## which binary numbers (of length 3) have a 1 at the 3rd position
out <- contains(3,3)
out
## check if true
bins <- lapply(out,to.binom,3)
bins
all(lapply(bins,`[`,3) == 1)
}
\author{
Florian Klinglmueller
}

