% Generated by roxygen2 (4.0.0): do not edit by hand
\name{FLIgmcP}
\alias{FLIgmcP}
\title{graph based multiple comparsions procedure using permutation tests}
\usage{
FLIgmcP(G, pvalues, ...)
}
\arguments{
\item{p}{unadjusted p-values}

\item{G}{graphMCP}

\item{...}{additional arguments to \code{\link{flip::npc}}}
}
\value{
vector with adjustde p-values
}
\description{
Perform a permutation test using adjusted p-values from a graph-based closed test procedure
}
\author{
Florian Klinglmueller
}

