% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_test.R
\name{from.binom}
\alias{from.binom}
\title{Binary vector to decimal number}
\usage{
from.binom(binom)
}
\arguments{
\item{binom}{}
}
\value{
Integer
}
\description{
turn a binary vector into a decimal number
}
\author{
Florian Klinglmueller
}

