% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gMCfliP}
\alias{gMCfliP}
\title{graph based multiple comparsions procedure using permutation tests}
\usage{
gMCfliP(permTP, G, ...)
}
\arguments{
\item{permTP}{flip.object}

\item{G}{graphMCP}

\item{...}{additional arguments to \code{\link{flip::npc}}}
}
\value{
vector with adjustde p-values
}
\description{
Perform a graph-based closed test procedure using permutation tests
}
\author{
Florian Klinglmueller
}

