% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.R
\name{join_to_row}
\alias{join_to_row}
\title{Join two matrices by row}
\usage{
join_to_row(c1, c2)
}
\arguments{
\item{c1}{a n1 \times m1 matrix}

\item{c2}{a n2 \times m2 matrix}
}
\value{
n1 + n2 \times m1*m2 matrix
}
\description{
Joins two matrices repeats each row of the first matrix by the
number of rows of the second matrix and column binds them together to a
}
\author{
Florian Klinglmueller
}

