% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_pw_contrasts}
\alias{make_pw_contrasts}
\title{Contrasts for many-to-one comparisons}
\usage{
make_pw_contrasts(g, control = 0)
}
\arguments{
\item{g}{Vector of treatment assignments}
}
\description{
Make contrast matrix for all (pairwise) many-to-one comparisons
}
\author{
Florian Klinglmueller
}

