% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{neuroscience}
\alias{neuroscience}
\title{Data from Cutini et al. (2014)}
\format{a \code{data.frame}}
\description{
Functional near infrared spectoscopy data from Cutini et al. (2014)
}
\details{
Data from 11 participants that performed a sequence of enumeration tasks. They were shown images with a number of dots ranging from 2 to 8, after which they quickly had to enter the number of dots they saw. The working hypothesis is that up to 4 dots dots the brain counts, but with more that  4 dots a combination of counting and estimation is performed.

Data were measured using multi-channel near-infrared spectosopy with 20 channels (10 left and 10 right hemisphere). Data consist of measurements from 11 patients participating in around 150 enumeration tasks each.
}
\author{
Florian Klinglmueller
}

