% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive.R
\name{normal_CER}
\alias{normal_CER}
\title{z - Test Conditional Error Rate}
\usage{
normal_CER(x1, g1, n, alpha = 0.025, sigma = 1, one_sample = F)
}
\arguments{
\item{x1}{First stage observations}

\item{g1}{First stage treatment assignments (1 for control, -1 for treatment)}

\item{n}{Preplanned overal sample size}

\item{alpha}{pre-fixed significance level}
}
\value{
numeric value of the conditional error rate
}
\description{
Computes the conditional type I error rate of a pre-planned z-test in a two-stage parallel group adaptive design. We condition on the observed first stage data and treatment assignment as well as the pre-planned second stage sample size.
}
\author{
Florian Klinglmueller
}

