% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.R
\name{omega}
\alias{omega}
\title{Reference space}
\usage{
omega(g1, g2 = NULL, g3 = NULL, restricted = TRUE, B = 1000)
}
\arguments{
\item{g1}{First stage treatment assignments}

\item{g2}{Second stage treatment assignments}

\item{g3}{Third stage treatment assignments}

\item{restricted}{Whether group sizes are considered fixed}

\item{B}{Number of permutations to be used (if smaller than all permutations)}
}
\value{
integer matrix
}
\description{
Reference/permutation space of a two-stage adaptive permutation test
}
\author{
Florian Klinglmueller
}

