% Generated by roxygen2 (4.0.0): do not edit by hand
\name{perm_test}
\alias{perm_test}
\title{perform permutation test}
\usage{
perm_test(x1, x2, g1, g2, stat, B, x3 = NULL, g3 = NULL)
}
\arguments{
\item{x1}{First stage observations}

\item{x2}{Second stage observations}

\item{g1}{First stage group assignments}

\item{g2}{Second stage group assignments}

\item{stat}{Function that computes test statistic}

\item{B}{Number of permutations to use}

\item{x3}{Third stage observations}

\item{g3}{Third stage group assignments}
}
\value{
p-value of the permutation test
}
\description{
Perform a permutation test (stratified by stages)
}
\author{
Florian Klinglmueller
}

