% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive.R
\name{permutation_CER}
\alias{permutation_CER}
\title{Permutation Conditional Error Rate}
\usage{
permutation_CER(x1, g1, x2, stat = sumdiff, permutations = 1000,
  alpha = 0.025, g2 = rep(c(-1, 1), each = length(x2)/2),
  one_sample = FALSE, restricted = TRUE, ...)
}
\arguments{
\item{x1}{vector of preplanned first stage observations}

\item{g1}{vector of first stage treatment assignments}

\item{x2}{vector of preplanned second stage observations}

\item{stat}{function computing the test statistic (see Details)}

\item{permutations}{number of permutations (rerandomizations) used to compute unconditional and conditional permutation distributions}

\item{alpha}{pre-fixed significance level}

\item{g2}{template vector for second stage treatment assignments}

\item{...}{additional options to \code{stat}}

\item{each}{}
}
\value{
numeric value of the conditional error rate
}
\description{
Computes the conditional type I error rate of a pre-planned permutation test in a two-stage adaptive design. We condition on the observed first stage data and treatment assignment as well as the observed second stage data - which we assume are obtained when the experiment reaches its preplanned sample size.
}
\details{
Based on the first stage data and treatment assignments one may perform sample size reassassment - and possibly other trial modifications - as long as the (preplanned) second stage sample size is not reduced.

\code{stat} needs to be a function of the form \code{function(x,g,...)} returning a numeric of length one. Possible options are \code{\link{sumdiff}}, \code{\link{meandiff}}, \code{\link{zstat}}

For the moment, we assume that observations are randomized using random allocation blocked by stages, (i.e. we resample using \code{sample (g1)}). \code{g2} does not have to be the actual second stage treatment assignments but just one possible example randomization, that fixes the treatment group sizes.
}
\author{
Florian Klinglmueller
}

