% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{power.z.test}
\alias{power.z.test}
\title{sample size formula one-sample z-test}
\usage{
power.z.test(power, delta, sd, sig.level = 0.025)
}
\arguments{
\item{power}{desired power}

\item{delta}{mean}

\item{sd}{standard deviation}

\item{sig.level}{significance level}
}
\value{
sample size
}
\author{
Florian Klinglmueller
}

