% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/statistics.R
\name{pw_zstat}
\alias{pw_zstat}
\title{Pairwise (many-to-one) z-scores}
\usage{
pw_zstat(x, g, control = 0, sigma = 1)
}
\arguments{
\item{x}{either a numeric vector or column matrix of outcomes (see Details)}

\item{g}{group assignments (see Details)}

\item{control}{Label that defines the control group}
}
\value{
numeric vector of test statistics
}
\description{
Matrix computation of all pairwise z-scores between (treatment) groups and one control group.
}
\details{
Group assignments can be passed via \code{g} using either a numeric/integer factor, vector, or matrix. All elements with values below or equal to zero are treated as belonging to the control, all above as belonging to the treatment group.

If either x or g are passed as matrices test statistics each computed for each column of outcomes or treatment assignments.  Passing both as matrices does not work!
}
\author{
Florian Klinglmueller
}

