% Generated by roxygen2 (4.0.0): do not edit by hand
\name{random_assignments}
\alias{random_assignments}
\title{Random combinations of two group assignments}
\usage{
random_assignments(n, B)
}
\arguments{
\item{n}{number of observations}

\item{B}{number of random combinations}
}
\value{
logical matrix
}
\description{
Returns a logical matrix with B random assignments (e.g. of patients) of \code{n} objects to two groups (e.g. treatment and control)
}
\author{
Florian Klinglmueller
}

