% Generated by roxygen2 (4.0.0): do not edit by hand
\name{random_reassignments}
\alias{random_reassignments}
\title{Random reassignments to two groups}
\usage{
random_reassignments(n, k, B)
}
\arguments{
\item{n}{number of observations}

\item{k}{number of observations in (e.g. treatment) group}

\item{B}{number of random combinations}
}
\value{
logical matrix
}
\description{
Returns a logical matrix with B random assignments (e.g. of patients) of \code{n} objects to two groups (e.g. treatment and control) given that \code{k} objects are in one group
}
\author{
Florian Klinglmueller
}

