% Generated by roxygen2 (4.0.0): do not edit by hand
\name{recycle}
\alias{recycle}
\title{recycle}
\usage{
recycle(a, b)
}
\arguments{
\item{a}{vector to be recycled}

\item{b}{vector to be matched}
}
\value{
vector of length at least \code{b}
}
\description{
recycle vector to match the length of another
}
\author{
Florian Klinglmueller
}

