% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{rnorm_cont}
\alias{rnorm_cont}
\title{Generate contaminated normally distributed data}
\usage{
rnorm_cont(n, mean = 0, sd = 1, cprop = 0.1, cshift = 3, csd = sd)
}
\arguments{
\item{n}{number of observations}

\item{mean}{mean value}

\item{sd}{standard deviation}

\item{cprop}{proportion of contaminated samples}

\item{cshift}{mean shift of contaminated samples}

\item{csd}{standard deviation of contaminated samples}
}
\value{
vector with \code{n} pseudo random numbers
}
\description{
Simulate from a contaminated normal distribution. A proportion \code{csd} of random samples are scaled by \code{csd} and randomly shifted to the left or right by \code{cshift}.
}
\author{
Florian Klinglmueller
}

