% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rnorm_hetero}
\alias{rnorm_hetero}
\title{Generate heteroskedastic data}
\usage{
rnorm_hetero(n, df = 1, ncp = 0)
}
\arguments{
\item{n}{number of observations}

\item{df}{degrees of freedom for the variance distribution}

\item{delta}{non-centrality parameter}
}

