% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{rnorm_scont}
\alias{rnorm_scont}
\title{Generate contaminated normally distributed data}
\usage{
rnorm_scont(n, mean = 0, sd = 1, cprop = 0.1, cmean = 3, csd = sd)
}
\arguments{
\item{n}{number of observations}

\item{mean}{mean value}

\item{sd}{standard deviation}

\item{cprop}{proportion of contaminated samples}

\item{cmean}{standard deviation of contaminated samples}
}
\value{
vector with \code{n} pseudo random numbers
}
\description{
Simulate from a location shift contaminated normal distribution
}
\author{
Florian Klinglmueller
}

