% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rnorm_tcont}
\alias{rnorm_tcont}
\title{Generate contaminated normally distributed data}
\usage{
rnorm_tcont(n, mean = 0, sd = 1, cprop = 0.1, csd = 3)
}
\arguments{
\item{n}{number of observations}

\item{mean}{mean value}

\item{sd}{standard deviation}

\item{cprop}{proportion of contaminated samples}

\item{csd}{standard deviation of contaminated samples}
}
\value{
vector with \code{n} pseudo random numbers
}
\description{
Simulate from a tail contaminated normal distribution
}
\author{
Florian Klinglmueller
}

