% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{simulate_trials}
\alias{simulate_trials}
\title{Simulate trials}
\usage{
simulate_trials(MCMC, n1, n, n3 = 0, r = 1/2, rfs = rnorm, rss = rnorm,
  ncp = 0, cond = c("both", "first", "none"), restricted = TRUE)
}
\arguments{
\item{MCMC}{Number of trials to simulate}

\item{n1}{First stage sample size}

\item{n}{Total preplanned sample size}

\item{n3}{Third stage sample size (maybe a function of \code{x},\code{g2})}

\item{rfs}{First stage random number generator}

\item{rss}{Second stage random number generator}

\item{ncp}{Non-centrality parameter}

\item{cond}{Condition on first and second stage observations ("both"), only on first stage observations ("first"), or unconditional of observations}
}
\value{
list object with simulated data
}
\description{
Simulate trial data
}
\author{
Florian Klinglmueller
}

