% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation.R
\name{strat_reassignments}
\alias{strat_reassignments}
\title{Stagewise two-group assignments}
\usage{
strat_reassignments(ns, ks, restricted = TRUE, B = NULL)
}
\arguments{
\item{ns}{vector of number of observations per stage}

\item{ks}{vector of number of observations in group (e.g. treatment) per stage}

\item{restricted}{}

\item{B}{number of assignments}
}
\value{
logical matrix of dimension \code{sum(ns)} \times \code{prod(choose(ns,ks))}
}
\description{
Returns all possible assignments to two groups stratified by stages.
}
\details{
Number of observations \code{ns} and observations from one group \code{ks} will be recycled to match the longer argument.
}
\author{
Florian Klinglmueller
}

