% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{sumdiff}
\alias{sumdiff}
\title{Sum of differences}
\usage{
sumdiff(x, g, ...)
}
\arguments{
\item{x}{either a numeric vector or column matrix of outcomes (see Details)}

\item{g}{group assignments (see Details)}
}
\value{
numeric vector of test statistics
}
\description{
Matrix computation of difference of sums between two groups
}
\details{
Group assignments can be passed via \code{g} using either a numeric/integer factor, vector, or matrix. All elements with values below or equal to zero are treated as belonging to the control, all above as belonging to the treatment group. 

If either x or g are passed as matrices test statistics each computed for each column of outcomes or treatment assignments.  Passing both as matrices does not work!
}
\author{
Florian Klinglmueller
}

