% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive.R
\name{t_test}
\alias{t_test}
\title{t-test p-value}
\usage{
t_test(x1, x2, g1, g2, x3 = NULL, g3 = NULL)
}
\arguments{
\item{x1}{First stage observations}

\item{x2}{Second stage observations}

\item{g1}{First stage treatment assignments}

\item{g2}{Second stage treatment assignments}

\item{x3}{Third stage observations}

\item{g3}{Third stage treatment assignments}
}
\value{
p-value of the t-test
}
\description{
Computes the p-value of the t-test (assuming equal variances).
}
\author{
Florian Klinglmueller
}

