% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to.binom}
\alias{to.binom}
\title{Convert to binary}
\usage{
to.binom(int, n = floor(log2(int)) + 1)
}
\arguments{
\item{int}{integer}

\item{n}{length of binary vector}
}
\value{
binary vector
}
\description{
make a binary vector out of a number
}
\examples{
to.binom(3,4)
}
\author{
Florian Klinglmueller
}

