% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adaptive.R
\name{z_test}
\alias{z_test}
\title{z-test p-value}
\usage{
z_test(x1, x2, g1, g2, sigma = 1, x3 = NULL, g3 = NULL)
}
\arguments{
\item{x1}{First stage observations}

\item{x2}{Second stage observations}

\item{g1}{First stage treatment assignments}

\item{g2}{Second stage treatment assignments}

\item{sigma}{Common standard deviation}

\item{x3}{Third stage observations}

\item{g3}{Third stage treatment assignments}
}
\value{
p-value of the z-test
}
\description{
Computes the p-value of the z-test (assuming known variances).
}
\author{
Florian Klinglmueller
}

