% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{adaptive_invnormtest_os}
\alias{adaptive_invnormtest_os}
\title{Inverse normal adaptive t-test}
\usage{
adaptive_invnormtest_os(x, n1, n, ne, alpha = 0.025)
}
\arguments{
\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{alpha}{}
}
\value{
Logical whether the test rejects
}
\description{
Adaptive combination test of stage-wise t-tests using the inverse normal combination function.
}
\author{
Florian Klinglmueller
}

