% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{adaptive_permtest_os}
\alias{adaptive_permtest_os}
\title{One-sample adaptive permutation test}
\usage{
adaptive_permtest_os(x, n1, n, ne, test_statistic, perms = 50000,
  alpha = 0.025)
}
\arguments{
\item{x}{Observations}

\item{n1}{First stage sample size}

\item{n}{Pre-planned total sample size}

\item{ne}{Extended sample size}

\item{perms}{Maximum number of permutations to use when computing permutation p-values and conditional error rates}

\item{alpha}{}

\item{combination_function}{Function to combine stage-wise (permutation) p-values}
}
\value{
Logical whether the test rejects
}
\description{
Adaptive permutation test one-sample problems
}
\author{
Florian Klinglmueller
}

