% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{all_assignments}
\alias{all_assignments}
\title{All assignments}
\usage{
all_assignments(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
integer matrix of size \code{n} x \code{2^n}
}
\description{
Returns a logical matrix with all possible assignments (e.g. of patients) of \code{n} objects to two groups (e.g. treatment and control)
}
\author{
Florian Klinglmueller
}

