% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLIgmcP.R
\name{FLIgmcP}
\alias{FLIgmcP}
\title{graph based multiple comparsions procedure using permutation tests}
\usage{
FLIgmcP(G, pvalues, ...)
}
\arguments{
\item{G}{graphMCP}

\item{...}{additional arguments to \code{\link{flip::npc}}}

\item{p}{unadjusted p-values}
}
\value{
vector with adjustde p-values
}
\description{
Perform a permutation test using adjusted p-values from a graph-based closed test procedure
}
\author{
Florian Klinglmueller
}

