% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{recycle}
\alias{recycle}
\title{recycle}
\usage{
recycle(a, b)
}
\arguments{
\item{a}{vector to be recycled}

\item{b}{vector to be matched}
}
\value{
vector of length at least \code{b}
}
\description{
recycle vector to match the length of another
}
\author{
Florian Klinglmueller
}

