% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{resamplingMCP-package}
\alias{resamplingMCP-package}
\title{resampling MCP

Fixed sample and adaptive permutation tests for single and multiple hypotheses

Provides functions to perform permutation tests.}
\description{
resampling MCP

Fixed sample and adaptive permutation tests for single and multiple hypotheses

Provides functions to perform permutation tests.
}

