% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resamplingMCP-package.R
\docType{package}
\name{resamplingMCP-package}
\alias{resamplingMCP-package}
\title{resampling MCP

Fixed sample and adaptive permutation tests for single and multiple hypotheses

Provides functions to perform permutation tests.}
\description{
resampling MCP

Fixed sample and adaptive permutation tests for single and multiple hypotheses

Provides functions to perform permutation tests.
}

