% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{rnorm_hetero}
\alias{rnorm_hetero}
\title{Generate heteroskedastic data}
\usage{
rnorm_hetero(n, df = 1, ncp = 0)
}
\arguments{
\item{n}{number of observations}

\item{df}{degrees of freedom for the variance distribution}

\item{delta}{non-centrality parameter}
}
\description{
Simulate normal heteroscedastic data with scaling of the variance drawn from a chi-square distribution with \
}

